Modifications needed for EnergyViewer
git-svn-id: https://svn.ee.ucl.ac.uk/repos/jefferson/userspacerouter@1275 83772449-cf8d-4b79-a1b1-a7e526131002

+++ usr/common/LocalHostInfo.java
@@ -210,7 +210,7 @@ public class LocalHostInfo {
 	}
 
 	// returns current energy consumption from the energy model
-	public double GetCurrentEnergyConsumption (float averageCPULoad, float averageIdleCPU, int memoryUsed, int freeMemory, long networkOutboundBytes, long networkIncomingBytes) {
+	public double GetCurrentEnergyConsumption (float averageCPULoad, float averageIdleCPU, float memoryUsed, float freeMemory, long networkOutboundBytes, long networkIncomingBytes) {
 		return energyModel.CurrentEnergyConsumption (averageCPULoad, averageIdleCPU, memoryUsed, freeMemory, networkOutboundBytes, networkIncomingBytes);
 	}
